// src/store/reducers/address/types/province.ts
export interface Province {
  ProvinceID?: number
  ProvinceName?: string
  CountryID?: number
  Code?: string
  NameExtension?: string[]
  IsEnable?: number
  RegionID?: number
  UpdatedBy?: number
  CreatedAt?: string
  UpdatedAt?: string
  CanUpdateCOD?: boolean
  Status?: number
  UpdatedIP?: string
  UpdatedEmployee?: number
  UpdatedSource?: string
  UpdatedDate?: string
}